OMG I was in a bug hole for 2 hours straight with a bad if()-statement. So burnt. Anywho, TraverserMapReduce is no longer needed -- everything is completely handled by GraphComputer.Memory. There is one last step to get to work in the model -- AggregateStep. The problem, it is both a CollectingBarrierStep AND a SideEffectStep. dah. I think I will get it -- just tired after the nonsense bug hunt. Opened up yet more TestSuite tests to ProcessComputerTestSuite.
